requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
bcrypt==4.1.2
blinker==1.8.2
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.2.0
Flask==3.0.3
Flask-Cors==4.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pydantic==2.7.1
pydantic_core==2.18.2
PyJWT==2.8.0
pymongo==4.8.0
python-dotenv==1.0.1
typing_extensions==4.12.2
uuid==1.30
validators==0.33.0
Werkzeug==3.0.3
gunicorn==20.0.4